+++ packages/stellar-wallet-snap/src/services/price/price-api/PriceApiClient.ts
@@ -13,20 +13,8 @@ import {
   InvalidHttpResponseException,
   normalizeHttpException,
 } from '../../../utils/errors';
-import type {
-  FiatExchangeRatesResponse,
-  GetHistoricalPricesParams,
-  GetHistoricalPricesResponse,
-  SpotPricesResponse,
-  VsCurrencyParam,
-} from './api';
-import {
-  FiatExchangeRatesResponseStruct,
-  GetHistoricalPricesParamsStruct,
-  GetHistoricalPricesResponseStruct,
-  GetSpotPricesParamsStruct,
-  GetSpotPricesResponseStruct,
-} from './api';
+import type { SpotPricesResponse, VsCurrencyParam } from './api';
+import { GetSpotPricesParamsStruct, GetSpotPricesResponseStruct } from './api';
 import { PriceApiException } from './exceptions';
 
 export class PriceApiClient {
@@ -48,31 +36,6 @@ export class PriceApiClient {
     this.#baseUrl = baseUrl;
   }
 
-  async getFiatExchangeRates(): Promise<FiatExchangeRatesResponse> {
-    try {
-      const url = buildUrl({
-        baseUrl: this.#baseUrl,
-        path: '/v1/exchange-rates/fiat',
-      });
-
-      const response = await this.#fetch(url);
-
-      if (!response.ok) {
-        throw new HttpResponseException(response.status);
-      }
-
-      const data = await response.json();
-      assertHttpResponse(data, FiatExchangeRatesResponseStruct);
-
-      return data;
-    } catch (error: unknown) {
-      return this.#throwError({
-        error,
-        fallbackError: 'Error fetching fiat exchange rates',
-      });
-    }
-  }
-
   async getSpotPrices(
     assetIds: CaipAssetType[],
     vsCurrency: VsCurrencyParam | string = 'usd',
@@ -92,7 +55,6 @@ export class PriceApiClient {
         queryParams: {
           vsCurrency,
           assetIds: assetIds.join(','),
-          includeMarketData: 'true',
         },
       });
 
@@ -110,63 +72,6 @@ export class PriceApiClient {
       return this.#throwError({
         error,
         fallbackError: 'Error fetching spot prices',
-      });
-    }
-  }
-
-  /**
-   * Business logic for `getHistoricalPrices`.
-   *
-   * @param params - The parameters for the request.
-   * @param params.assetType - The asset type of the token.
-   * @param params.timePeriod - The time period for the historical prices.
-   * @param params.from - The start date for the historical prices.
-   * @param params.to - The end date for the historical prices.
-   * @param params.vsCurrency - The currency to convert the prices to.
-   * @returns The historical prices for the token.
-   * @throws {HttpResponseException} When the HTTP response status is not successful.
-   * @throws {InvalidHttpResponseException} When the response body is invalid.
-   * @throws {PriceApiException} When the request fails for another reason.
-   */
-  async getHistoricalPrices(
-    params: GetHistoricalPricesParams,
-  ): Promise<GetHistoricalPricesResponse> {
-    try {
-      assertHttpRequestParams(params, GetHistoricalPricesParamsStruct);
-
-      const url = buildUrl({
-        baseUrl: this.#baseUrl,
-        path: '/v3/historical-prices/{assetType}',
-        pathParams: {
-          assetType: params.assetType,
-        },
-        queryParams: {
-          ...(params.timePeriod !== undefined && {
-            timePeriod: params.timePeriod,
-          }),
-          ...(params.from !== undefined && { from: params.from.toString() }),
-          ...(params.to !== undefined && { to: params.to.toString() }),
-          ...(params.vsCurrency !== undefined && {
-            vsCurrency: params.vsCurrency,
-          }),
-        },
-        encodePathParams: false,
-      });
-
-      const response = await this.#fetch(url);
-
-      if (!response.ok) {
-        throw new HttpResponseException(response.status);
-      }
-
-      const historicalPrices = await response.json();
-      assertHttpResponse(historicalPrices, GetHistoricalPricesResponseStruct);
-
-      return historicalPrices;
-    } catch (error: unknown) {
-      return this.#throwError({
-        error,
-        fallbackError: 'Error fetching historical prices',
       });
     }
   }